Explicitly delete the node object bound to the Machine on deletion
In the latest version of the kubernetes the node object is not
automatically removed once the underlying VM is deleted. Thus
this leaves the system with additional nodes that are in `NotReady`
state for which the VMs have already been removed.
Ultimately, this should be handled in the k/k controllers. However,
for the time being the cluster-api controller can handle the node
removal upon machine deletion

Issue #256
Change-Id: Ie365aa915b2a08e36065f88ad12389ff7b75120c

(cherry picked from commit 78624f0a2e02b2bfc44b49ae7e5bfc6495de3e9f)