NKS-3154 Remove finalizer if parent cluster is being deleted (#51)
* NKS-3154 Remove finalizer if parent cluster is being deleted

The deletion timestamp tends to be slow to arrive for the vspherecluster, and when it does the event doesn't always come in.

With this change we remove the finalizer if the parent cluster (CAPI) is being deleted, which will cause the vsphere cluster to disappear immediately.

* Add "tagged" annotation

Retagging is very, very slow. And it's done on every single reconcile!